requirements.txt
asgiref==3.4.1
attrs==19.3.0
backports.entry-points-selectable==1.1.0
cachelib==0.1.1
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cssselect==1.1.0
cssutils==2.2.0
distlib==0.3.2
dnspython==2.1.0
filelock==3.0.12
Flask==2.0.0
Flask-Session==0.3.2
idna==2.10
itsdangerous==2.0.0
jellyfish==0.8.2
Jinja2==3.0.0
lib50==1.1.11
lxml==4.6.3
markdown2==2.4.0
MarkupSafe==2.0.0
nodejs==0.1.1
optional-django==0.1.0
pexpect==4.8.0
platformdirs==2.2.0
premailer==3.8.0
ptyprocess==0.7.0
py3-validate-email==1.0.0
python-decouple==3.4
python-dotenv==0.17.1
python-vlc==3.0.12118
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
submit50==3.0.0
termcolor==1.1.0
urllib3==1.26.4
virtualenv==20.7.0
Werkzeug==2.0.0
